requirements.txt
Django~=5.0
python-dotenv~=1.0.0
djangorestframework~=3.14.0